pyproject.toml
[tool.poetry]
name = "x-app-web"
version = "0.1.0"
description = "x-app-web"
authors = ["juan-mendes-DEV <jp436807@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.4"
factory-boy = "^3.3.1"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-debug-toolbar = "^5.0.1"
django-jazzmin = "^3.0.1"
djangorestframework-simplejwt = "^5.4.0"
psycopg2-binary = "^2.9.10"
django-environ = "^0.12.0"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
